Refactoring architecture
* move/rename scripts
